requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
Flask-Session==0.5.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
